Functions for managing the cluster are moved to a separate package
