+++ cli/cluster/lua/get_membership_instances_body.lua
