+++ cli/cluster/lua/probe_instances_body.lua
